Password Checker
The objective is to create a program to check if your password is secure. It will check the length of the password and if it contains numbers. In addition, if your password is equal to or less than seven characters long the program will tell you "Your password should be at least eight characters long" and then ask the user to type in their password again.