Nintendo eShop - Assignment 4
The objective is to create a program that acts as a shopping cart for Nintendo games. There will be an initial list of games the user can choose from and another list that will be the user’s “shopping cart.” The user must also have the function to take games already in the shopping cart and remove them. In addition, there must be a quit option and a start-over option.