Create 308CINFassignment5.py
The objective is to create a program that collects user information and stores it in a list. Note we also have to use dictionaries and tuples. To do this, I need to ask for user input on set categories, and then group related categories. For example, I could gather information on hair and eye color, and then group them into a tuple. A similar format of grouping alike categories could apply to dictionaries as well.