Nintendo Adventure Game (CINF 308 Assignment 6)
The objective is to create a text-based game that has multiple storylines and endings. The game must have the ability to change a level or score based on what the user does. In addition, lists, dictionaries, and tuples must be used. For this objective, I created a Nintendo Adventure Game! The game includes storylines from various popular gaming franchises. That includes Mario, Kirby, and the Legend of Zelda. Each location has a set difficulty, unique enemies, and a series of endings.