Create 308CINFassignment8.py
The objective is to take a past program created and make substantial improvements by incorporating objects. I picked my assignment two program, and plan to alter how the program determines the level of the user. The program will ask for the user’s time and gender they are competing in and assign it to a class. From there, the info will be put through a series of if and elif statements in a Determine_Level function.