Create 308CINFassignment9.py
The objective is to take last week’s program and create another source that uses objects but not inheritance. Last week’s program was the runner program that determines your skill level based on the gender you're competing in and time. I plan to add a “SprintTechnique” that asks if the runner/user believes they have good form. This will be done by using composition.