308CINF_Midterm
The objective is to take a past Python project and make substantial improvements to it. I plan to make adjustments to assignment five. The program will act as a customer relationship management system. You should be able to add new customers, delete customers, view customer details, view system logs, and be able to quit.