Create 308CINFassignment10.py
The objective is to use at least two methods we have learned this past week. This includes search, sort, insert, update, and delete. For this assignment, I created a program that acts as a list of a user’s favorite video games. The program will have a search and insert function for the user to add and look at games on the list.