CINF308_Assignment12
The objective is to find a company idea and create a program that will solve a substantial problem. I plan to solve business analytical issues by creating charts based on user input. This will allow the company to analyze potential trends and relationships to better comprehend their current economic situation.